+++ src/plugin.hpp
@@ -35,4 +35,5 @@ extern Model* modelMirror;
 extern Model* modelAffix;
 extern Model* modelAffixMicro;
 extern Model* modelGrip;
+extern Model* modelGlue;
 extern Model* modelAudioInterface64;